+++ crypto/src/main/rust/circuit/random/binaryuniformdistribution.rs
@@ -54,13 +54,7 @@ impl<'a, Z: IntegerModRing + Clone> BinaryUniformDistribution<'a, Z> {
     }
 
     fn to_bits(integer: &Z) -> Vec<Z> {
-        decompose_integer(
-            integer,
-            Z::Int::LIMB_ONE,
-            Z::Int::LIMB_ONE,
-            Z::BITS as usize,
-        )
-        .into()
+        decompose_integer(integer, Z::Int::LIMB_ONE, Z::Int::LIMB_ONE, Z::BITS).into()
     }
 }
